Address SSH URL review feedback
