+++ Sources/TabManager+CompatibilityTypes.swift
@@ -1,0 +1,54 @@
+import Foundation
+import CoreGraphics
+import Bonsplit
+
+// MARK: - Direction Types for Backwards Compatibility
+
+/// Split direction for backwards compatibility with old API
+enum SplitDirection {
+    case left, right, up, down
+
+    var isHorizontal: Bool {
+        self == .left || self == .right
+    }
+
+    var orientation: SplitOrientation {
+        isHorizontal ? .horizontal : .vertical
+    }
+
+    /// If true, insert the new pane on the "first" side (left/top).
+    /// If false, insert on the "second" side (right/bottom).
+    var insertFirst: Bool {
+        self == .left || self == .up
+    }
+}
+
+/// Resize direction for backwards compatibility
+enum ResizeDirection {
+    case left, right, up, down
+
+    var splitOrientation: String {
+        switch self {
+        case .left, .right:
+            return "horizontal"
+        case .up, .down:
+            return "vertical"
+        }
+    }
+
+    /// A split controls the target pane's right/bottom edge when the target is
+    /// the first child, and left/top edge when the target is the second child.
+    var requiresPaneInFirstChild: Bool {
+        switch self {
+        case .right, .down:
+            return true
+        case .left, .up:
+            return false
+        }
+    }
+
+    /// Positive values move the divider toward the second child (right/down).
+    var dividerDeltaSign: CGFloat {
+        requiresPaneInFirstChild ? 1 : -1
+    }
+}
